LPD-88102 Extract PortletDataHandler types and rename fixture
