+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/js/components/forms/content_selector/ContentSelector.tsx
@@ -6,8 +6,8 @@
 import ClayAlert from '@clayui/alert';
 import React from 'react';
 
+import {PortletDataHandlerSection} from '../../../types/portletDataHandlerSection';
 import {updateSelection} from '../../../utils/contentSelection';
-import {PortletDataHandlerSection} from '../../../utils/mockPortletDataHandlerSections';
 import ContentSection, {SectionSelection} from './ContentSection';
 
 export type ContentSelection = Record<string, SectionSelection>;
