+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/js/components/forms/formik/FormikFieldContentSelector.tsx
@@ -6,7 +6,7 @@
 import {useField, useFormikContext} from 'formik';
 import React from 'react';
 
-import {PortletDataHandlerSection} from '../../../utils/mockPortletDataHandlerSections';
+import {PortletDataHandlerSection} from '../../../types/portletDataHandlerSection';
 import ContentSelector, {
 	ContentSelection,
 } from '../content_selector/ContentSelector';
