+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/js/utils/contentSelection.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {PortletDataHandlerControl} from './mockPortletDataHandlerSections';
+import {PortletDataHandlerControl} from '../types/portletDataHandlerSection';
 
 export type HandlerSelection =
 	| {
