+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/js/utils/mockExportPreview.ts
@@ -3,35 +3,10 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export interface PortletDataHandlerBoolean {
-	label: string;
-	name: string;
-	portletDataHandlerControls?: PortletDataHandlerControl[];
-	type: 'Boolean';
-}
-
-export interface PortletDataHandlerChoice {
-	choices: {label: string; name: string}[];
-	label: string;
-	name: string;
-	type: 'Choice';
-}
-
-export type PortletDataHandlerControl =
-	| PortletDataHandlerBoolean
-	| PortletDataHandlerChoice;
-
-export type PortletDataHandler = Omit<PortletDataHandlerBoolean, 'type'>;
-
-export interface PortletDataHandlerSection {
-	label: string;
-	name: string;
-	portletDataHandlers: PortletDataHandler[];
-}
-
-export interface ExportPreview {
-	portletDataHandlerSections: PortletDataHandlerSection[];
-}
+import {
+	ExportPreview,
+	PortletDataHandlerSection,
+} from '../types/portletDataHandlerSection';
 
 const sections: PortletDataHandlerSection[] = [
 	{
